Danger Mode preferences integration into preferences repository (#324)
* feat: Add confirmation button

The button will be used for the call or sms confirmation but it is fully modular so it can be used for other thing as needed.

* feat: Add confirmation button

The button will be used for the call or sms confirmation but it is fully modular so it can be used for other thing as needed.

* fix: remove preview for line coverage

* feat: Add automatic preferences to repository local and remote