Automatic dangermode complete (#327)
* feat: Add confirmation button

The button will be used for the call or sms confirmation but it is fully modular so it can be used for other thing as needed.

* feat: Add confirmation button

The button will be used for the call or sms confirmation but it is fully modular so it can be used for other thing as needed.

* fix: remove preview for line coverage

* feat: Add automatic preferences to repository local and remote

* feat: Add exclusive options

Add exclusive options for call and sms.

* feat: Add robustness to DangerModeCardViewModel.kt

Add robustness against crash

* fix: Fix check for active hazard in orchestrator

Fix useless check for manual activation of danger mode.

* fix: Fix tests not working in DangerModeCardTest.kt

Fix flacky tests and danger mode view model nto working properly

* fix: Fix flaky test in DangerModeCardViewModelTest.kt

Fix DangerModeCardViewModel.kt because it made a test flaky with Jacoco.

* fix: add requested changes

* fix: fix test not working in DangerModeCardTest.kt

* fix: fix test not working in DangerModeCardViewModelTest.kt